Move bulk method from own Interface to IndexerInterface
