+++ packages/seal-memory-adapter/src/MemoryIndexer.php
@@ -13,14 +13,13 @@ declare(strict_types=1);
 
 namespace Schranz\Search\SEAL\Adapter\Memory;
 
-use Schranz\Search\SEAL\Adapter\BulkableIndexerInterface;
 use Schranz\Search\SEAL\Adapter\IndexerInterface;
 use Schranz\Search\SEAL\Marshaller\Marshaller;
 use Schranz\Search\SEAL\Schema\Index;
 use Schranz\Search\SEAL\Task\SyncTask;
 use Schranz\Search\SEAL\Task\TaskInterface;
 
-final class MemoryIndexer implements IndexerInterface, BulkableIndexerInterface
+final class MemoryIndexer implements IndexerInterface
 {
     private readonly Marshaller $marshaller;
