+++ packages/seal-redisearch-adapter/src/RediSearchIndexer.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Schranz\Search\SEAL\Adapter\RediSearch;
 
-use Schranz\Search\SEAL\Adapter\BulkableIndexerInterface;
 use Schranz\Search\SEAL\Adapter\BulkHelper;
 use Schranz\Search\SEAL\Adapter\IndexerInterface;
 use Schranz\Search\SEAL\Marshaller\Marshaller;
@@ -21,7 +20,7 @@ use Schranz\Search\SEAL\Schema\Index;
 use Schranz\Search\SEAL\Task\SyncTask;
 use Schranz\Search\SEAL\Task\TaskInterface;
 
-final class RediSearchIndexer implements IndexerInterface, BulkableIndexerInterface
+final class RediSearchIndexer implements IndexerInterface
 {
     private readonly Marshaller $marshaller;
